Removed dataEncipherment keyUsage
dataEncipherment is not applicable on practical uses. 

For a browser client digitalSignature and keyAgreement are the ones to use.
For a SMIME email client, digitalSignature and keyEncipherment are the ones to choose.
For PKCS #7/AdES/PDF (endorsement, not deniable) document signature. Now this been should be renamed contentCommitment (same sense, just different name)